requirements.txt
pandas
tweepy
selenium
textblob
finnhub-python
matplotlib
torch
transformers
scipy
yfinance
requests
scikit-learn
plotly
google-cloud-storage